remove pop-shared and reactive-search
