+++ apps/production/package.json
@@ -2,7 +2,6 @@
   "name": "pop_production",
   "version": "1.0.3",
   "dependencies": {
-    "@appbaseio/reactivesearch": "2.16.1",
     "autosize": "^4.0.2",
     "bootstrap": "^4.3.1",
     "classnames": "^2.2.5",
@@ -18,7 +17,6 @@
     "isomorphic-fetch": "^2.2.1",
     "jszip": "^3.2.1",
     "leaflet": "^1.3.1",
-    "pop-shared": "^0.18.0",
     "prop-types": "^15.7.2",
     "qs": "^6.7.0",
     "rc-steps": "^3.1.1",
